requirements.txt
bcrypt==4.2.0
certifi==2024.8.30
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.3.2
cryptography==43.0.1
distlib==0.3.8
filelock==3.16.1
identify==2.6.1
idna==3.10
iniconfig==2.0.0
jmespath==1.0.1
nodeenv==1.9.1
packaging==24.1
paramiko==3.5.0
platformdirs==4.3.6
pluggy==1.5.0
pre-commit==3.8.0
pycparser==2.22
PyNaCl==1.5.0
pytest==8.3.3
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
scp==0.15.0
six==1.16.0
urllib3==2.2.3
virtualenv==20.26.6
